completed category, event and order routes, and one product route, other two product routes skipped because one has complicated querying and other needs aws s3 configuration, which will be added in dedicated commits later

+++ src/main/java/com/devrish/martcart/MartcartApplication.java
@@ -103,7 +103,7 @@ public class MartcartApplication {
 					.address("123 XYZ Street, India")
 					.quantity(2L)
 					.productId(sampleProduct)
-					.totalPrice((long)(sampleProduct.getCurrentPrice() * 2))
+					.totalPrice(sampleProduct.getCurrentPrice() * 2)
 					.userId(sampleBuyer)
 					.build();
